refactor: split index.ts into crawler/types/utils files
